VSCode: Move `isScarbProject` logic to separate file
commit-id:f5c8f336
