+++ vscode-cairo/src/scarbProject.ts
@@ -1,0 +1,28 @@
+import * as fs from "fs";
+import * as path from "path";
+import * as vscode from "vscode";
+
+export async function isScarbProject(): Promise<boolean> {
+  const depth = 20;
+  const workspaceFolders = vscode.workspace.workspaceFolders;
+  if (
+    !!workspaceFolders?.[0] &&
+    (await isScarbProjectAt(path.dirname(workspaceFolders[0].uri.path), depth))
+  )
+    return true;
+  const editor = vscode.window.activeTextEditor;
+  return (
+    !!editor &&
+    (await isScarbProjectAt(path.dirname(editor.document.uri.path), depth))
+  );
+}
+
+async function isScarbProjectAt(path: string, depth: number): Promise<boolean> {
+  if (depth == 0) return false;
+  const isFile = await fs.promises
+    .access(path + "/Scarb.toml", fs.constants.F_OK)
+    .then(() => true)
+    .catch(() => false);
+  if (isFile) return true;
+  return isScarbProjectAt(path + "/..", depth - 1);
+}
